requirements-dev.txt
autoflake
autopep8
flake8
isort
pytest
pytest-cov
pytest-django
